add README.md

+++ Features/steps/PageObjects/ShoppingCartPageObject.py
@@ -1,3 +1,4 @@
+from decimal import *
 
 
 class ShoppingCartPageObject:
@@ -14,10 +15,10 @@ class ShoppingCartPageObject:
         sum_items = 0
         for item in self.items_list:
             # need to remove currency before the price
-            sum_items += float(item.find_element_by_xpath('./div/div[2]/p/span').text[1:])
+            sum_items += Decimal(item.find_element_by_xpath('./div/div[2]/p/span').text[1:])
         return sum_items
 
     def get_subtotal(self):
         # need to remove space and currency before the price
-        return float(self.subtotal.text.strip()[1:])
+        return Decimal(self.subtotal.text.strip()[1:])
